requirements.txt
transformers>=4.21.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0
